+++ lib/zaikio/hub/machine.rb
@@ -10,6 +10,10 @@ module Zaikio
       attributes :name, :updated_at, :created_at, :site_id, :kind, :capabilities,
                  :manufacturer_id, :serial_number, :manufacturer, :owner_id
 
+      # Associations
+      has_many :availabilities, class_name: "Zaikio::Hub::Availability",
+                                uri: "machines/:machine_id/availabilities"
+
       def self.kinds
         %w[sheetfed_digital_press web_digital_press digital_press sheetfed_offset_press
            web_offset_press offset_press cutter shape_cutter folder box_maker