+++ lib/zaikio/hub/specialist.rb
@@ -9,6 +9,10 @@ module Zaikio
       # Attributes
       attributes :name, :updated_at, :created_at, :site_id, :kind, :owner_id, :capabilities
 
+      # Associations
+      has_many :availabilities, class_name: "Zaikio::Hub::Availability",
+                                uri: "specialists/:specialist_id/availabilities"
+
       def self.kinds
         %w[shipping other boxing].freeze
       end