+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Support `availabilities` for `Machine` and `Specialist`
+
 ## [0.15.3]
 
 - Improve german and english translations